# swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "version":"1.0.0",
        "description":"API for Postaway Social Media Project",
        "title":"Postaway API"
    },
    "servers":[
        {
            "url":"http://localhost:3000"
        }
    ],
    "components":{
        "securitySchemes":{
            "JWT":{
                "in":"header",
                "name":"Authorization",
                "type":"apikey"
            }
        }
    },
    "paths":{
...
        "/api/posts/":{
            "get":{
              "tags": ["Posts"],
              "summary": "Get posts",
              "description": "Get posts based on user credentials",
              "security": [{ "JWT": {} }],
              "parameters":[
                {
                    "name":"userId",
                    "in":"path",
                    "required":false,
                    "schema":{
                        "type":"string"
                    },
                    "description":"filter by userId"
                }
              ],
              "responses":{
                "500": { "description": "Server error" },
                "200": { "description": "OK" }
              }
            },

            "post":{
              "tags": ["Posts"],
              "summary": "Create post",
              "description": "Create a new post",
              "security": [{ "JWT": {} }],
                "requestBody":{
                    "required":true,
                    "content":{
                        "multipart/form-data":{
                            "schema":{
                                "type":"object",
                                "required":["caption","image"],
                                "properties":{
                                    "caption":{ "type":"string" },
                                    "image":{"type":"string","format":"binary"}
                                }
                            }
                        }
                    }
                },
              "responses":{
                "400":{"description":"Unauthorized"},
                "500": { "description": "Server error" },
                "201": { 
                    "description": "Post is created",
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "properties":{
                                    "message":{"type":"string"},
                                    "post":{
                                        "type":"object",
                                        "properties":{
                                            "id":{"type":"integer"},
                                            "userId":{"type":"integer"},
                                            "caption":{"type":"string"},
                                            "imageUrl":{"type":"string"}
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
              }
            }
        },
...
    }
}